requirements.txt
asgiref==3.3.1
certifi==2020.6.20
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.2
cycler==0.10.0
defusedxml==0.7.0rc2
Django==3.1.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
et-xmlfile==1.0.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
kiwisolver==1.3.0
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.1.0
openpyxl==3.0.6
pandas==1.1.3
Pillow==8.0.1
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
